moved core.errors to exceptions
