+++ bugzoo/exceptions.py
